PERF-750: remove reference to jira-soke-tests-eu-s3 bucket
